Reduce the size of the Pawn Hash Zobrist to short from long.

+++ EubosChess/src/main/java/eubos/position/IPositionAccessors.java
@@ -12,7 +12,7 @@ public interface IPositionAccessors {
 	public int getMoveNumber();
 	public int getPlyNumber();
 	public long getHash();
-	public int getPawnHash();
+	public short getPawnHash();
 	public String getFen();
 	public boolean isThreefoldRepetitionPossible();
 	public boolean isInsufficientMaterial();
